+++ CHANGELOG.rst
@@ -9,6 +9,11 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 `Unreleased`_
 -------------
 
+Fixed
+.....
+
+- Properly handle I2C write errors
+
 `0.4.0`_ 2021-01-20
 -------------------